requirements.txt
# Airflow 3.0 providers
apache-airflow-providers-fab>=2.0.0
apache-airflow-providers-standard>=0.0.1

# Core pipeline dependencies
pyspark==3.5.4
pandas==2.2.3
pyarrow==18.1.0

# Testing
pytest==8.3.4

# Linting (optional, for candidate use)
ruff==0.8.6
